requirements.txt
Django~=3.2.9
djangorestframework~=3.12.4
django-filter~=21.1
drf-yasg~=1.20.0
django-cors-headers~=3.10.0
djangorestframework-simplejwt~=5.0.0
